+++ insight/admin/__init__.py
@@ -1,0 +1,3 @@
+from .base_admin import *  # noqa: F403
+from .user_weekly_trend_admin import *  # noqa: F403
+from .weekly_trend_admin import *  # noqa: F403
