+++ noti/models.py
@@ -54,7 +54,7 @@ class NotiMailLog(TimeStampedModel):
 
     user = models.ForeignKey(
         VelogUser,
-        on_delete=models.CASCADE,
+        on_delete=models.SET_NULL,
         related_name="email_logs",
         verbose_name="수신자",
     )
